ai-service/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==11.0.0
numpy==2.1.2
gunicorn==21.2.0
torch==2.4.1
transformers==4.46.1
safetensors==0.4.5